-Wformat error in SharedBuffer::tryCreateArrayBuffer https://bugs.webkit.org/show_bug.cgi?id=195004
Reviewed by Darin Adler.

Seems C++ has no format specifier appropriate for printing the result of sizeof. We should
just not try to print it. Anyway, that's easy in this case, because sizeof(char) is
guaranteed to be 1. This code was an attempt to be pedantic to account for mythical systems
with char larger than one byte, but perhaps it didn't realize sizeof always returns
multiples of char and so sizeof(char) is always one even on such mythical systems.

Note the sizeof(char) use two lines up is left since it's not clear that switching it to 1
would actually be more readable.

* platform/SharedBuffer.cpp:
(WebCore::SharedBuffer::tryCreateArrayBuffer const):


Identifier: 209596@main
git-svn-id: https://svn.webkit.org/repository/webkit/trunk@242327 268f45cc-cd09-0410-ab3c-d52691b4dbfc