Make sure an empty host matches the internal representation "nullOrigin" in WebCore::RegistrableDomain::matches() https://bugs.webkit.org/show_bug.cgi?id=195435
Reviewed by Brent Fulgham.

No new tests. Tests are currently failing and will pass again with
this patch.

WebCore::RegistrableDomain::matches() is a quick way to compare a
RegistrableDomain with a URL. Since RegistrableDomain represents the
empty host as "nullOrigin," a URL with an empty host needs to match
that representation in WebCore::RegistrableDomain::matches().
Failure to do so caused debug assertions in fast/ layout tests after
https://trac.webkit.org/changeset/242603/webkit.

* platform/RegistrableDomain.h:
(WebCore::RegistrableDomain::matches const):


Identifier: 209758@main
git-svn-id: https://svn.webkit.org/repository/webkit/trunk@242620 268f45cc-cd09-0410-ab3c-d52691b4dbfc